# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/prescriptions": {
      "post": {
        "operationId": "Prescriptions_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrescriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prescription created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request — invalid payload (DTO validation failed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Doctor role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity — payload is well-formed but the referenced patientId does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create a new prescription (Doctor Only)",
        "tags": [
          "Prescriptions"
        ]
      },
      "get": {
        "operationId": "Prescriptions_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página (1-indexed).",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Cantidad de items por página.",
            "schema": {
              "minimum": 1,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtrar por estado de prescripción.",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "CONSUMED"
              ]
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "Filtrar desde fecha (ISO 8601).",
            "schema": {
              "example": "2026-01-01",
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "Filtrar hasta fecha (ISO 8601).",
            "schema": {
              "example": "2026-12-31",
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Case-insensitive substring match against prescription notes and item names. Role-based visibility is preserved.",
            "schema": {
              "maxLength": 200,
              "example": "amoxi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns paginated list of prescriptions based on user role",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedResultDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/PrescriptionResponseDto"
                          }
                        },
                        "meta": {
                          "$ref": "#/components/schemas/PaginationMetaDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request — invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List prescriptions (Paginated). Role-scoped: Doctor sees own authored, Patient sees own received, Admin sees all.",
        "tags": [
          "Prescriptions"
        ]
      }
    },
    "/prescriptions/{id}": {
      "get": {
        "operationId": "Prescriptions_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Prescription ID (UUID v4)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the prescription detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — you do not have access to this prescription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found — prescription does not exist or access forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get prescription detail by ID",
        "tags": [
          "Prescriptions"
        ]
      }
    },
    "/prescriptions/{id}/consume": {
      "patch": {
        "operationId": "Prescriptions_markAsConsumed",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Prescription ID (UUID v4)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumePrescriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prescription status updated to CONSUMED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — Patient role required or prescription does not belong to patient",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found — prescription not found or does not belong to patient",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflict — prescription has already been consumed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Mark prescription as consumed (Patient Only)",
        "tags": [
          "Prescriptions"
        ]
      }
    },
    "/prescriptions/{id}/pdf": {
      "get": {
        "operationId": "Prescriptions_downloadPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Prescription ID (UUID v4)",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Streams the generated PDF file",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden — you do not have access to this prescription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found — prescription not found or not authorized to download",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Download prescription PDF",
        "tags": [
          "Prescriptions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Prescription Management API",
    "description": "API documentation for the MVP Prescription Management System.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
...
      "Role": {
        "type": "string",
        "enum": [
          "ADMIN",
          "DOCTOR",
          "PATIENT"
        ],
        "description": "User role for RBAC"
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code of the error"
          },
          "message": {
            "type": "string",
            "example": "Invalid JSON payload",
            "description": "Error message describing what went wrong"
          },
          "path": {
            "type": "string",
            "example": "/auth/login",
            "description": "The request path that caused the error"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-13T10:30:00.000Z",
            "description": "ISO 8601 timestamp when the error occurred"
          }
        },
        "required": [
          "statusCode",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current page number (1-indexed)"
          },
          "limit": {
            "type": "number",
            "example": 10,
            "description": "Number of items per page"
          },
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total number of items across all pages"
          },
          "totalPages": {
            "type": "number",
            "example": 5,
            "description": "Total number of pages"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "PaginatedResultDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "PrescriptionStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONSUMED"
        ],
        "description": "Current status of the prescription"
      },
      "PrescriptionItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Prescription item unique identifier",
            "example": "aaaa1111-bbbb-2222-cccc-3333dddd4444",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Amoxicillin",
            "description": "Name of the medication"
          },
          "dosage": {
            "type": "string",
            "example": "500mg",
            "description": "Dosage of the medication",
            "nullable": true
          },
          "quantity": {
            "type": "number",
            "example": 30,
            "description": "Quantity to dispense (number of units)",
            "nullable": true
          },
          "instructions": {
            "type": "string",
            "example": "Take 1 pill every 8 hours",
            "description": "Instructions for the patient",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "PrescriptionUserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "user@clinic.com",
            "format": "email"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "PrescriptionAuthorSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Doctor unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "specialty": {
            "type": "string",
            "example": "Cardiology",
            "description": "Medical specialty",
            "nullable": true
          },
          "medicalId": {
            "type": "string",
            "example": "MED-12345",
            "description": "Medical license / registration number",
            "nullable": true
          },
          "signatureText": {
            "type": "string",
            "example": "Dr. Jane Doe",
            "description": "Text rendering of the doctor signature",
            "nullable": true
          },
          "signatureImageUrl": {
            "type": "string",
            "example": "https://cdn.clinic.com/signatures/jane-doe.png",
            "description": "URL to the doctor signature image",
            "format": "uri",
            "nullable": true
          },
          "user": {
            "description": "Doctor account",
            "allOf": [
              {
                "$ref": "#/components/schemas/PrescriptionUserSummaryDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "user"
        ]
      },
      "PrescriptionPatientSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Patient unique identifier",
            "example": "456e7890-e89b-12d3-a456-426614174001",
            "format": "uuid"
          },
          "birthDate": {
            "format": "date-time",
            "type": "string",
            "example": "1990-05-21T00:00:00.000Z",
            "description": "Patient date of birth (ISO-8601 timestamp)",
            "nullable": true
          },
          "user": {
            "description": "Patient account",
            "allOf": [
              {
                "$ref": "#/components/schemas/PrescriptionUserSummaryDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "user"
        ]
      },
      "PrescriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Prescription unique identifier",
            "example": "789e0123-e89b-12d3-a456-426614174002",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "description": "Short URL-safe code used by the QR on the PDF",
            "example": "RX-A1B2C3D4E5"
          },
          "status": {
            "description": "Current status of the prescription",
            "example": "PENDING",
            "allOf": [
              {
                "$ref": "#/components/schemas/PrescriptionStatus"
              }
            ]
          },
          "items": {
            "description": "List of medications in the prescription",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrescriptionItemResponseDto"
            }
          },
          "notes": {
            "type": "string",
            "example": "Follow up in 2 weeks",
            "description": "Additional notes from the doctor",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the prescription was created",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the prescription was last updated",
            "example": "2026-01-15T12:30:00.000Z"
          },
          "consumedAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the prescription was consumed",
            "example": "2026-01-16T08:45:00.000Z",
            "nullable": true
          },
          "authorId": {
            "type": "string",
            "description": "Author (Doctor) unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "description": "Patient unique identifier",
            "example": "456e7890-e89b-12d3-a456-426614174001",
            "format": "uuid"
          },
          "author": {
            "description": "Prescribing doctor summary",
            "allOf": [
              {
                "$ref": "#/components/schemas/PrescriptionAuthorSummaryDto"
              }
            ]
          },
          "patient": {
            "description": "Patient summary",
            "allOf": [
              {
                "$ref": "#/components/schemas/PrescriptionPatientSummaryDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "code",
          "status",
          "items",
          "createdAt",
          "updatedAt",
          "authorId",
          "patientId",
          "author",
          "patient"
        ]
      },
      "PrescriptionItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Amoxicillin",
            "description": "Name of the medication"
          },
          "dosage": {
            "type": "string",
            "example": "500mg",
            "description": "Dosage of the medication"
          },
          "quantity": {
            "type": "number",
            "example": 30,
            "description": "Quantity to dispense (number of units)"
          },
          "instructions": {
            "type": "string",
            "example": "Take 1 pill every 8 hours",
            "description": "Instructions for the patient"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreatePrescriptionDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "Patient identifier. Accepts either the Patient profile id (Patient.id) or the related User.id — the server resolves either form. See GET /users/patients which exposes both.",
            "format": "uuid"
          },
          "items": {
            "description": "List of medications in the prescription",
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PrescriptionItemDto"
            }
          },
          "notes": {
            "type": "string",
            "example": "Follow up in 2 weeks",
            "description": "Additional notes from the doctor"
          }
        },
        "required": [
          "patientId",
          "items"
        ]
      },
      "ConsumePrescriptionDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Picked up at the pharmacy on May 13",
            "description": "Optional free-text reason persisted in the prescription audit log.",
            "maxLength": 500
          }
        }
      },
...
    }
  }
}